+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -4,10 +4,10 @@ CHANGELOG
 5.1.0
 -----
 
- * Marked `MicroKernelTrait::configureRoutes()` as `@internal` and `@final`.
- * Deprecated not overriding `MicroKernelTrait::configureRouting()`.
+ * Made `MicroKernelTrait::configureContainer()` compatible with `ContainerConfigurator`
  * Added a new `mailer.message_bus` option to configure or disable the message bus to use to send mails.
  * Added flex-compatible default implementations for `MicroKernelTrait::registerBundles()` and `getProjectDir()`
+ * Deprecated passing a `RouteCollectionBuiler` to `MicroKernelTrait::configureRoutes()`, type-hint `RoutingConfigurator` instead
 
 5.0.0
 -----